fixed compilation errors

+++ blockend/facets/ViewFacet.sol
@@ -14,10 +14,10 @@ contract ViewFacet is Modifiers {
 
     function positionHealthFactorFutures(IUniswapV3Pool pool, address owner, uint24 contractId) external view returns (bool isLiquidable, int256 actualMarginPercentage, int256 actualMarginAmount) {
         Types.PerpFuture memory future = s.futureRecord[address(pool)][owner][contractId];
-        int256 priceDiffInPercentage = int256(((TWAPOracle.getPoolTWAP(pool.token0(), pool.token1(), pool.fee(), future.collateralAmount, true) / future.collateralAmount) * future.initialPrice) / future.initialPrice);
-        actualMarginPercentage = future.futureType == Types.FutureType.CALL ? priceDiffInPercentage * future.leverage / Constants.BASIS_POINTS : -priceDiffInPercentage * future.leverage / Constants.BASIS_POINTS;
+        int256 priceDiffInPercentage = int256(((TWAPOracle.getPoolTWAP(pool.token0(), pool.token1(), pool.fee(), uint128(future.collateralAmount), true) / future.collateralAmount) * future.initialPrice) / future.initialPrice);
+        actualMarginPercentage = future.futureType == Types.FutureType.CALL ? priceDiffInPercentage * int256(future.leverage) / int24(Constants.BASIS_POINTS) : -priceDiffInPercentage * int256(future.leverage) / int24(Constants.BASIS_POINTS);
         // 5% * 10_000 * 10 * 100 / (10_000) = 50 (profits)
-        actualMarginAmount = (actualMarginPercentage * future.leverage * future.collateralAmount) / Constants.BASIS_POINTS;
+        actualMarginAmount = (actualMarginPercentage * int256(future.leverage) * int256(future.collateralAmount)) / int24(Constants.BASIS_POINTS);
         if (actualMarginPercentage < -int256(future.maintenanceMargin)) isLiquidable = true;
     }
 }
